JobsPilotAI/requirements.txt
streamlit
requests
beautifulsoup4
lxml
pandas
SQLAlchemy
python-dotenv
PyYAML
pdfplumber
scikit-learn
Pillow
APScheduler
